refactor: remove up contracts-node and refactor the functionality into a function
